--- NOAA-River-Report.bat
@@ -1,3 +1,0 @@
-@echo off 
-python3 app\main.py 
-PAUSE
